+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added: `pluck` function which is curried
 - Added: `prop` function which is curried
 - Added: `every` function which is curried
+- Added: `any` function which is curried
 - Fixed: `filter` functionality to make more sense and not use a makeshift (bad) reduce
 - Fixed: removed `Object.entries` from `filter` didn't realize babel wasn't able to transpile it like I hoped